Split component and component_id out of glue.core.data, and updated imports. Also replaced almost all absolute imports by relative imports.
