+++ glue/clients/scatter_client.py
@@ -5,9 +5,9 @@ from functools import partial
 import numpy as np
 
 from ..core.client import Client
-from ..core.data import Data, IncompatibleAttribute, ComponentID, CategoricalComponent
+from ..core.data import Data, IncompatibleAttribute, ComponentID
 from ..core.subset import RoiSubsetState, RangeSubsetState, CategoricalRoiSubsetState, AndState
-from ..core.roi import PolygonalROI, RangeROI, CategoricalRoi, RectangularROI
+from ..core.roi import PolygonalROI, RangeROI, RectangularROI
 from ..core.util import relim
 from ..core.edit_subset_mode import EditSubsetMode
 from ..core.message import ComponentReplacedMessage
