+++ glue/core/data_factories/tests/test_data_factories.py
@@ -8,7 +8,8 @@ import numpy as np
 from numpy.testing import assert_allclose, assert_array_equal
 
 from ... import data_factories as df
-from ...data import CategoricalComponent, Data
+from ...data import Data
+from ...component import CategoricalComponent
 from ...tests.util import make_file
 
 from ....tests.helpers import (requires_astropy, requires_astropy_ge_03,
